Update / export types
